Initial commit
This has been slowly developed for a while now, since 2018, but since
testing it required a lot trial and error with Travis-CI, which made git
history a mess, it was done in a separate repository. Now that all the
features I wanted to implement are done and there are no apparent bugs,
it's time to publish it in a proper repo for everyone else to use.