Use TRAVIS_*_WEB_URL environment variables
Travis-CI provides TRAVIS_JOB_WEB_URL and TRAVIS_BUILD_WEB_URL variables
that are set to URLs of the current job and the build it belongs to.

We can use those variables for job/build URLs for the default release
body messages, instead of having to construct those URLs ourselves.

We can also use those variables to figure out if we are running on the
.org or the .com instance, removing the need for --travis-instance-org
and --travis-instance-com arguments.

This would also allow us to remove the awkward --travis-instance-custom
argument and change it into --travis-api-url to match --github-api-url.