Don't hardcode version in the install script
Keeping that version information up to date is annoying as it makes the
release process two-step: you have to create a bdist release, which
doesn't include the scripts, then calculate sha of it, update the
install script and only then create a sdist release, which does include
the scripts.
It would also be impossible to pull off if we were one day to include
those scripts into a bdist release. It's the same issue as creating a
file that contains its own hash -- you need to know file's final
contents in order to calculate its hash, but the contents are not final
until they include its hash.