Allow restricting builds in Latest Release check
Latest Release checks if the current build is the latest for the branch,
and if it's not, it doesn't update the Latest Release, assuming that the
latest build runs CIRP too, so it will be the one to update it. However,
this assumption is not always valid, e.g. a user could have setup only
cron and api builds to run CIRP, so if the latest build is not a cron or
api build, we have no business assuming that the latest build will run
CIRP. There is no way for the script to know which event types to check
for and which to ignore, so a user-provided option is required here.