Add license statement to the helper scripts
These scripts are intended to be added into codebases of other projects,
so explicitly including the license statement into them makes it a bit
easier for those projects to comply with the license, as it requires
the copyright and the license text to be included.

I was initially against adding the license text to those as it blows up
the 10-line scripts, making them seem more than what they actually are,
but having integrated CIRP into several other projects as well as seeing
other projects integrate it on their own, with those projects adding the
license text to the scripts, made me realize that there is more value to
having the license text be included than I had initially thought.